+++ src/lib.rs
@@ -3,8 +3,10 @@ pub mod auto_config;
 pub mod cli;
 pub mod config;
 pub mod core;
+pub mod target;
 pub mod ui;
 pub mod utils;
+pub mod wrapper;
 
 #[cfg(feature = "self-update")]
 pub mod updater;